Extract common code into a separate file
This is a bit of pain when writing an Ansible module, because they have
to be self-contained files.

For now, I've come up with a simple Makefile that uses `cat` to prepend
the contents of `common.py` to each module.

A better solution would be to put everything we need from common.py into
the 'pygerrit' module, which is already an external dependency.