gerrit_group: Allow setting included_groups
This lets you include one group in another. So for example you can have
a 'Submitters' group that includes the 'Registered Users' group, to
allow all registered users to be submitters.

Also, the module now returns more than just the group_info structure
returned by Gerrit. The Gerrit REST API defines what the group_info
structure can contain and, to be consistent with that, we shoudn't add
included_groups information to it directly.