Allow specifying Gerrit URL and admin credentials in the environment
This saves you having to specify credentials for each task in Ansible
playbook. It's inconvenient if you're running the playbook with a remote
host, rather than localhost, but I'm not sure why you would do that.

Also, rename gerrit_username to gerrit_admin_username and
gerrit_password to gerrit_admin_password. This makes it clearer when
using the 'gerrit_account' module whose username is whose.