gerrit_account: Ignore failure to remove an account from a group
This is a kludge to work around the fact that the Gerrit
accounts/xx/groups API call returns groups that the user can't actually
be removed from. If a user is in group A only due to their membership in
group B which includes group A, we can't remove the user from group A
or we'll get a 404 error.